+++ RunBubbleShooter.sh
@@ -1,4 +1,0 @@
-cd examples/bubbleshooter
-npm install
-npm run build
-npm run dev