+++ RunColorBlockJam.sh
@@ -1,4 +1,0 @@
-cd examples/colorblockjam
-npm install
-npm run build
-npm run dev