+++ RunHelloWorld.sh
@@ -1,4 +1,0 @@
-cd examples/helloworld
-npm install
-npm run build
-npm run dev