+++ RunHexasort.sh
@@ -1,4 +1,0 @@
-cd examples/hexasort
-npm install
-npm run build
-npm run dev