+++ RunMatch3.sh
@@ -1,4 +1,0 @@
-cd examples/match3
-npm install
-npm run build
-npm run dev