+++ RunSolitaire.sh
@@ -1,4 +1,0 @@
-cd examples/solitaire
-npm install
-npm run build
-npm run dev