+++ RunTemplate.sh
@@ -1,4 +1,0 @@
-cd templates/example_template
-npm install
-npm run build
-npm run dev