+++ examples/avoidance/src/AvoidanceUIIds.ts
@@ -1,4 +1,0 @@
-export enum AvoidanceUIIds {
-  GameScreen   = "Avoidance.GameScreen",
-  GameOverPopup = "Avoidance.GameOverPopup",
-}