+++ examples/avoidance/src/views/IGameScreenView.ts
@@ -1,3 +1,0 @@
-import type { IScreenView } from "@gamebyte/gamelabsjs";
-
-export type IGameScreenView = IScreenView;