+++ examples/bubbleshooter/src/models/IScore.ts
@@ -1,7 +1,0 @@
-import { InjectionToken } from "@gamebyte/gamelabsjs";
-
-export interface IScore {
-  readonly value: number;
-}
-
-export const IScore = new InjectionToken<IScore>("IScore");