+++ examples/helloworld/src/HelloWorldUIIds.ts
@@ -1,3 +1,0 @@
-export enum HelloWorldUIIds {
-  GameScreen = "HelloWorld.GameScreen",
-}