+++ examples/match3/src/Match3UIIds.ts
@@ -1,3 +1,0 @@
-export enum Match3UIIds {
-  GameScreen = "Match3.GameScreen",
-}