+++ examples/solitaire/src/SolitaireUIIds.ts
@@ -1,3 +1,0 @@
-export enum SolitaireUIIds {
-  GameScreen = "Solitaire.GameScreen",
-}