+++ examples/solitaire/src/constants/SlotType.ts
@@ -1,6 +1,0 @@
-export enum SlotType {
-  Stock = "Stock",
-  Waste = "Waste",
-  Foundation = "Foundation",
-  Tableau = "Tableau",
-}