+++ examples/uiplayground/src/UIPlaygroundUIIds.ts
@@ -1,3 +1,0 @@
-export enum UIPlaygroundUIIds {
-  PlaygroundShell = "UIPlayground.PlaygroundShell",
-}