+++ examples/watersort/src/WaterSortAssetIds.ts
@@ -1,6 +1,0 @@
-export enum WaterSortAssetIds {
-  Background  = "WaterSort.Background",
-  Bottle      = "WaterSort.Bottle",
-  BottleShine = "WaterSort.BottleShine",
-  Star        = "WaterSort.Star",
-}